requirements.txt
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
DAWG-Python==0.7.2
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.1
lxml_html_clean==0.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyaes==1.6.1
pyarrow==16.0.0
pyasn1==0.6.0
pydeck==0.9.0
Pygments==2.17.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy2-dicts-uk==2.4.1.1.1460299261
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
regex==2024.4.28
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.33.0
Telethon==1.35.0
tenacity==8.2.3
tinysegmenter==0.3
tldextract==5.1.2
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
